Adds in Sophronia Broadcasting as a Company
Sophronia Broadcasting is referenced by the Ingredient Beacon given to the Cook with, "Please enjoy your Sophronia Broadcasting's 'Plasteel Chef' Ingredients Box, exactly as shown in the hit show!", as well as the Curators Hero Beacon with, "Please enjoy your Sophronia Broadcasting's 'History Comes Alive branded' Costume Set, exactly as shown in the hit show!". I think it'd be a sensible addition, and adds nicely to any any reporter crewmembers, or a reporter AI who don't know what company they should say they work for. The original idea behind the company was that they were involved in mind control experiments, but it wasn't referenced in the game so I didn't add it.